Cleaned up relationship between API and Encounter component and tests.
